pyproject.toml
[project]
name = "test-radar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-stubs>=5.2.0",
    "djangorestframework>=3.16.0",
    "httpx>=0.28.1",
    "pytest>=8.4.0",
]

[project.entry-points.pytest11]
pytest_radar = "pytest_radar.plugin"

[tool.setuptools.packages.find]
where = ["."]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
    "flake8-aaa>=0.17.0",
    "mypy>=1.16.0",
    "ondivi>=0.7.2",
    "pre-commit>=4.2.0",
    "ruff>=0.11.13",
    "wemake-python-styleguide>=1.1.0",
]
test = [
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "server.settings"
# radar_endpoint = 'http://127.0.0.1:8000'

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN101",
  "ANN401",
  "D1",
  "D203",
  "D213",
  "ERA001",
  "FIX002",
  "PGH003",
  "Q000",
  "RUF012",
  "RUF100",
  "S105",
  "S603",
  "S607",
  "TD002",
  "TD003",
  "UP022",
  "UP032"
]
exclude = [
  "manage.py",
  "*/migrations/*",
  "pytest_radar",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001", "FLY002", "S101"]
"pytest_radar/*" = ["ARG001", "EM101", "INP001", "T201"]

[tool.ruff.format]
quote-style = "single"

[tool.poetry.group.dev.dependencies]
django-stubs = {extras = ["compatible-mypy"], version = "^5.2.0"}

